requirements.txt
--extra-index-url=http://pypi.ionata.com/

Django==1.4.20
lxml
Pillow==2.4.0
django-ajax-validation==0.1.3
django-appconf==0.4.1
django-compressor==1.1.2
#django-debug-toolbar==0.8.5
django-email-confirmation==0.2
django-staticfiles==1.1.2
django-timezones==0.2
gunicorn==0.13.4
pytz==2011n
sorl-thumbnail==11.12
wsgiref==0.1.2
django-formwizard==1.0
pyparsing==1.5.7
django-blog-zinnia==0.10.1
django-singletons==0.1.6-1
django-ckeditor==3.6.3-ionata3
django-filemanager==0.1.0
django-reversion==1.7.1
django-positions==0.4.3
python-dateutil==2.1
django-facebook==4.2.2
django-registration==0.8
privatebeta==0.4.1
django-celery-with-redis==3.0
django-ajax-forms
redis
--allow-external pyodbc
--allow-unverified pyodbc
pyodbc==3.0.7
django-pyodbc-mssql==1.0-ionata1
django-chainable-manager==0.3.1
django-haystack==1.2.7
xapian-haystack==1.1.5beta
django-filter==0.5.4
django-nested-formsets==0.2.0
django-maintenancemode==0.10
docutils==0.9.1
fabric==1.5.1
Pygments==1.5
django-dbtemplates==1.3
micawber==0.2.5
South==0.7
PyYAML==3.10
Markdown==2.2.1
djangorestframework==2.3.5
requests
django-discover-runner
-e git+https://github.com/ionata/wadofstuff-django-serializers.git@v1.1.0#egg=wadofstuff-django-serializers
python-magic
django-extensions
